Breakout core modules into their own files
