+++ lib/socket/core/authenticate.ex
@@ -1,0 +1,164 @@
+defmodule Socket.Core.Authenticate do
+  @moduledoc """
+  Handles the "authenticate" event
+  """
+
+  require Logger
+
+  alias Gossip.Applications
+  alias Gossip.Channels
+  alias Gossip.Games
+  alias Gossip.Presence
+  alias Metrics.SocketInstrumenter
+  alias Socket.Backbone
+  alias Socket.Core
+  alias Socket.Games, as: SocketGames
+  alias Socket.Players
+  alias Socket.Tells
+
+  @disable_debug_seconds 300
+
+  def process(state, payload) do
+    SocketInstrumenter.connect()
+
+    with {:ok, game} <- validate_socket(payload),
+         {:ok, supports} <- validate_supports(payload) do
+      finalize_auth(state, game, payload, supports)
+    else
+      {:error, :invalid} ->
+        Logger.debug("Disconnecting - invalid authenticate")
+        SocketInstrumenter.connect_failure()
+
+        {:disconnect, %{event: "authenticate", status: "failure", error: "invalid credentials"}, state}
+
+      {:error, :missing_supports} ->
+        Logger.debug("Disconnecting - missing supports")
+        SocketInstrumenter.connect_failure()
+
+        {:disconnect, %{event: "authenticate", status: "failure", error: "missing supports"}, state}
+
+      {:error, :must_support_channels} ->
+        Logger.debug("Disconnecting - must support channels")
+        SocketInstrumenter.connect_failure()
+
+        {:disconnect, %{event: "authenticate", status: "failure", error: "must support channels"}, state}
+
+      {:error, :unknown_supports} ->
+        Logger.debug("Disconnecting - unknown set of supports")
+        SocketInstrumenter.connect_failure()
+
+        {:disconnect, %{event: "authenticate", status: "failure", error: "includes unknown supports"}, state}
+    end
+  end
+
+  def finalize_auth(state, game, payload, supports) do
+    channels = Map.get(payload, "channels", [])
+    players = Map.get(payload, "players", [])
+    debug = Map.get(payload, "debug", false)
+
+    state =
+      state
+      |> Map.put(:status, "active")
+      |> Map.put(:game, game)
+      |> Map.put(:supports, supports)
+      |> Map.put(:channels, channels)
+      |> Map.put(:players, players)
+      |> Map.put(:debug, debug)
+
+    listen_to_channels(channels)
+    Players.maybe_listen_to_players_channel(state)
+    SocketGames.maybe_listen_to_games_channel(state)
+    Tells.maybe_subscribe(state)
+    Backbone.maybe_finalize_authenticate(state)
+
+    maybe_schedule_disable_debug(state)
+
+    SocketInstrumenter.connect_success()
+    Logger.info("Authenticated #{game.name} - subscribed to #{inspect(channels)} - supports #{inspect(supports)}")
+    Presence.track(state)
+
+    response = %{
+      event: "authenticate",
+      status: "success",
+      payload: %{
+        unicode: "✔️",
+        version: Gossip.version(),
+      }
+    }
+
+    send(self(), :heartbeat)
+
+    {:ok, response, state}
+  end
+
+  defp maybe_schedule_disable_debug(state) do
+    case state.debug do
+      true ->
+        Process.send_after(self(), {:disable_debug}, :timer.seconds(@disable_debug_seconds))
+
+      false ->
+        :ok
+    end
+  end
+
+  defp validate_socket(payload) do
+    client_id = Map.get(payload, "client_id")
+    client_secret = Map.get(payload, "client_secret")
+
+    case Games.validate_socket(client_id, client_secret, payload) do
+      {:ok, game} ->
+        {:ok, game}
+
+      {:error, :invalid} ->
+        Applications.validate_socket(client_id, client_secret)
+    end
+  end
+
+  defp validate_supports(payload) do
+    with {:ok, supports} <- get_supports(payload),
+         {:ok, supports} <- check_supports_for_channels(supports),
+         {:ok, supports} <- check_unknown_supports(supports) do
+      {:ok, supports}
+    end
+  end
+
+  defp get_supports(payload) do
+    case Map.get(payload, "supports", :error) do
+      :error ->
+        {:error, :missing_supports}
+
+      [] ->
+        {:error, :missing_supports}
+
+      supports ->
+        {:ok, supports}
+    end
+  end
+
+  defp check_supports_for_channels(supports) do
+    case "channels" in supports do
+      true ->
+        {:ok, supports}
+
+      false ->
+        {:error, :must_support_channels}
+    end
+  end
+
+  defp check_unknown_supports(supports) do
+    case Enum.all?(supports, &Core.valid_support?/1) do
+      true ->
+        {:ok, supports}
+
+      false ->
+        {:error, :unknown_supports}
+    end
+  end
+
+  defp listen_to_channels(channels) do
+    channels
+    |> Enum.map(&Channels.ensure_channel/1)
+    |> Enum.each(&Core.subscribe_channel/1)
+  end
+end
+
